refactor: moved global upload function to ipfs.js.
