+++ src/views/Dashboard/Lab/Services/Detail.vue
@@ -85,7 +85,7 @@
 
 <script>
 import { mapGetters } from "vuex"
-import { upload } from "@/lib/ipfs/upload"
+import { upload } from "@/lib/ipfs"
 import { updateService, deleteService } from "@/lib/polkadotProvider/command/services"
 
 export default {
